Add qtdb_trace and qtdb_resume functions
Add qtdb_trace and qtdb_resume functions to cadnano/util.py.  These
functions make PDB usable by removing and restoring the PyQt input hook
when pausing and resuming program execution so that PDB can run.

If the standard 'import pdb; pdb.set_trace()' is called without removing
the input hook, the message
> QCoreApplication::exec: The event loop is already running
is spammed to the console.

Import and call qtdb_trace to start debugging at a certain point.  Note
that, with the current implementation of qtdb_trace, PDB starts in the
frame corresponding to the qtdb_trace function, so the developer will
have to hit 'n' once to get back to the correct frame.
>>> from cadnano.util import qtdb_trace, qtdb_resume
>>> ...
>>> qtdb_trace() # Stop program execution and start debugging here

Call qtdb_resume from the PDB prompt when ready to resume the program.
The developer will have to press 'c' a couple times to break out of the
PDB prompt as well.