Prompt the user to save changes post clicking 'New'
Prior to this change, clicking 'New' in the toolbar would not prompt the
user to save his or her changes, clobbering any work that the user
completed.

This change adds a check to prompt the user to save unchanged work
should the user choose to do so.

Note that currently, the environment variable 'CADNANO_DISCARD_UNSAVED'
is set to True in cadnanoqt.py.  Consequently, the feature being added
in this commit is currently unreachable (as the prompt to save unchanged
work will never be shown to the user when this env variable is set to
True).  The implementation of this feature has been tested by not
setting the environment variable to 'True', however.