2.5.1:  Remove legacysliceview directory
To ensure that comments and documentation updates that are made in the
sliceview directory are reflected in the legacysliceview, removing this
directory.  Changes to implement the legacysliceview will be made
directly in the sliceview directory, and once modifications are complete
the sliceview directory will be renamed to legacysliceview and a
checked-out version of the sliceview directory will be restored to the
sliceview directory.