Fix bug where floats were saved as strings
When values in the PropertyEditor were changed by typing in new values,
new values were saved as strings, which caused a crash when opening the
saved file.

This commit casts values of keys that should be floats to floats.
Looked into UI-side validation (since right now the editor allows
non-floats to be entered as a value) without much success; it seems that
a change in PropertyEditorWidget.py (and associated classes) would need
to be made.

Fixes #112