2.5.1:  Fix crashing when clicking outside of GridItem
Prior to this fix, clicking on the outer edge of the `GridItem`'s stroke
would cause a crash, as this was more than `_RADIUS` away from the
center of the `GridItem` and the `closest_point` code assumed that a
click would always be at most `_RADIUS` away from the center of a
GridItem.

In lieu of increasing `_RADIUS` when searching (and possibly returning
a location that is adjacent to the actual location (e.g. when clicking
on the edge between two `GridItem`s, fail back to a brute force nearest
neighbors search when the closest `GridItem` is more than `_RADIUS` away
from the clicked location.