2.5:  Fix the long-standing napart undo-redo bug
Fix the bug whereby creating a new lattice, undoing, and then clicking
either lattice button or redoing would crash cadnano.

```
Traceback (most recent call last):
  File "./cadnano/views/gridview/gridrootitem.py", line 112, in clearSelectionsSlot
    self.select_tool.deselectItems()
  File "./cadnano/views/gridview/tools/selectgridtool.py", line 237, in deselectItems
    group.clearSelectionRect()
  File "./cadnano/views/gridview/tools/selectgridtool.py", line 488, in clearSelectionRect
    self.removeFromGroup(bri)
RuntimeError: wrapped C/C++ object of type GridSelectionGroup has been deleted
Abort trap: 6
```

In `setPartItem` of `SelectGridTool`, the parent of `self.group`, a
`GridSelectionGroup`, would be set to the given parameter `part_item`.
This is problematic in the case of the bug, as the old part item would
be garbage collected upon clicking either of the lattice buttons, as
nothing would reference it anymore.  Because the parent of the
`GridSelectionGroup` is the now un-referenced `part_item`, it is garbage
collected.  However, `self.group`, the owning reference to the
`GridSelectionGroup` Python object, is still in scope and is therefore
not garbage collected.

This change detects when this happens, and – when it does – creates a
new `GridSelectionGroup`.