Implement copying and pasting in the SliceView
Enable copying and pasting in the SliceView.  When the select tool is
selected, copying a VH or group of VHs copies the VH(s) to a buffer and
shows a preview of what the pasted VHs will look like.

Pasting ensures that the VHs are pasted in the correct position for
their parity.  Furthermore, pasting (either via clicking or via the
context menu) is disabled if pasting would be illegal (i.e. if pasting
would result in a VH being placed on top of another VH).