Add support for auto-detection of lattice type
In lieu of storing the lattice type in the JSON file, programmatically
determine what the lattice type is.

In addition, update the legacy import code so that it positions virtual
helices in the correct position upon import.  Updates were also made to
the legacy export code to support exporting designs on a square lattice
for backward compatibility as well.